+++ miniRT/srcs/lib_math/vec3_op1.c
@@ -1,6 +1,6 @@
 #include "miniRT.h"
 
-t_vec3	add_vec3_and_const(t_vec3 vector, double k)
+t_vec3	add_vec3_and_const(t_vec3 vector, float k)
 {
 	t_vec3	new_vector;
 
@@ -10,7 +10,7 @@ t_vec3	add_vec3_and_const(t_vec3 vector, double k)
 	return (new_vector);
 }
 
-t_vec3	sub_vec3_and_const(t_vec3 vector, double k)
+t_vec3	sub_vec3_and_const(t_vec3 vector, float k)
 {
 	t_vec3	new_vector;
 
@@ -20,7 +20,7 @@ t_vec3	sub_vec3_and_const(t_vec3 vector, double k)
 	return (new_vector);
 }
 
-t_vec3	mul_vec3_and_const(t_vec3 vector, double k)
+t_vec3	mul_vec3_and_const(t_vec3 vector, float k)
 {
 	t_vec3	new_vector;
 
@@ -30,7 +30,7 @@ t_vec3	mul_vec3_and_const(t_vec3 vector, double k)
 	return (new_vector);
 }
 
-t_vec3	div_vec3_and_const(t_vec3 vector, double k)
+t_vec3	div_vec3_and_const(t_vec3 vector, float k)
 {
 	t_vec3	new_vector;
