+++ miniRT/srcs/raytracing/camera_init.c
@@ -1,6 +1,6 @@
 #include "miniRT.h"
 
-void	check_limit_angle(double *angle)
+void	check_limit_angle(float *angle)
 {
 	if (*angle > 89.0f)
 		*angle = 89.0f;
@@ -19,15 +19,15 @@ void	init_euler_angles(t_camera *camera)
 
 void	update_camera_ray(t_ray *cam_ray, t_data *data)
 {
-	double	ratio;
-	double	scale;
+	float	ratio;
+	float	scale;
 
-	ratio = (double)WIDTH / (double)HEIGHT;
+	ratio = (float)WIDTH / (float)HEIGHT;
 	scale = tan(data->scene->camera->fov * 0.5);
 	cam_ray->dir.coord[X] = (2.0 * ((data->pixel_x + 0.5)
-				/ (double)WIDTH) - 1.0) * ratio * scale;
+				/ (float)WIDTH) - 1.0) * ratio * scale;
 	cam_ray->dir.coord[Y] = (1 - 2 * ((data->pixel_y + 0.5)
-				/ (double)HEIGHT)) * scale;
+				/ (float)HEIGHT)) * scale;
 	cam_ray->dir.coord[Z] = -1.0;
 	cam_ray->dir = mul_dir_and_matrix4(cam_ray->dir, data->cam_to_world_matrix);
 	copy_vec3(&cam_ray->origin, *data->scene->camera->origin);
